Implement CRUD operations for Author, Book, and User entities in respective controllers and services

+++ library-spring-ws-rest/src/main/java/no/hvl/dat152/rest/ws/repository/AuthorRepository.java
@@ -3,7 +3,6 @@
  */
 package no.hvl.dat152.rest.ws.repository;
 
-
 import org.springframework.data.repository.CrudRepository;
 
 import no.hvl.dat152.rest.ws.model.Author;
@@ -11,7 +10,6 @@ import no.hvl.dat152.rest.ws.model.Author;
 /**
  * @author tdoy
  */
-public interface AuthorRepository extends CrudRepository<Author, Long> {
-	
+public interface AuthorRepository extends CrudRepository<Author, Integer> {
 
 }
